move viewer entities to viewer
