+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/entities/Role.kt
@@ -1,4 +1,4 @@
-package li.doerf.feeder.common.entities
+package li.doerf.feeder.viewer.entities
 
 import org.springframework.security.core.GrantedAuthority
