+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/security/JwtTokenProvider.kt
@@ -3,8 +3,8 @@ package li.doerf.feeder.viewer.security
 import io.jsonwebtoken.JwtException
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.SignatureAlgorithm
-import li.doerf.feeder.common.entities.Role
 import li.doerf.feeder.viewer.HttpException
+import li.doerf.feeder.viewer.entities.Role
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpStatus
