+++ viewer/src/main/kotlin/li/doerf/feeder/viewer/services/UserService.kt
@@ -1,11 +1,11 @@
 package li.doerf.feeder.viewer.services
 
-import li.doerf.feeder.common.entities.Role
-import li.doerf.feeder.common.entities.User
-import li.doerf.feeder.common.repositories.UserRepository
 import li.doerf.feeder.common.util.getLogger
 import li.doerf.feeder.viewer.HttpException
 import li.doerf.feeder.viewer.dto.UserResponseDto
+import li.doerf.feeder.viewer.entities.Role
+import li.doerf.feeder.viewer.entities.User
+import li.doerf.feeder.viewer.repositories.UserRepository
 import li.doerf.feeder.viewer.security.JwtTokenProvider
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
