+++ viewer/src/test/kotlin/li/doerf/feeder/viewer/controllers/TestHelper.kt
@@ -1,7 +1,7 @@
 package li.doerf.feeder.viewer.controllers
 
-import li.doerf.feeder.common.entities.User
-import li.doerf.feeder.common.repositories.UserRepository
+import li.doerf.feeder.viewer.entities.User
+import li.doerf.feeder.viewer.repositories.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
